skriver om createTable til react-komponent <Table/>
