+++ src/app/personside/infotabs/ytelser/pleiepenger/Pleiepengerperiode.tsx
@@ -4,8 +4,8 @@ import styled from 'styled-components';
 import YtelserPeriode from '../felles-styling/YtelserPeriode';
 import { Undertittel } from 'nav-frontend-typografi';
 import { formaterDato, NOKellerNull, prosentEllerNull } from '../../../../../utils/stringFormatting';
-import { StyledTable } from '../../../../../utils/tableUtils';
 import theme from '../../../../../styles/personOversiktTheme';
+import { StyledTable } from '../../../../../utils/table/StyledTable';
 
 interface Props {
     periode: Pleiepengeperiode;
