+++ src/utils/tableUtils.test.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react';
 
 import { shallow } from 'enzyme';
-import { createTable } from './tableUtils';
+import { Table } from './table/Table';
 
 test('lager tabell basert på input', () => {
     const header: Array<string> = ['kolonne 1', 'kolonne 2'];
     const body: Array<Array<string>> = [['rad1 kolonne 1', 'rad1 kolonne 2']];
 
-    const result = shallow(createTable(header, body));
+    const result = shallow(<Table tittelRekke={header} rows={body} />);
     const expectedResult = (
         <table>
             <thead>
